fix bug 6432 split the old MAKE_NJOBS concept into the following formula make jobs is: (length $DISTCC_HOSTS)*$JOBS_PER_MAKE+$EXTRA_JOBS) this gives you greater freedom such that when distcc is disabled the number of make jobs can be scaled back. EXTRA_JOBS is stored as MAKE_NJOBS to maintain compatibility, JOBS_PER_MAKE defaults to 0.
This also includes a hack to workaround the fact that
DISTCC_HOSTS="" doesnt actually work in all cases, doing so
now works in all cases and a new non-hacky method (DISABLE_DISTCC=yes)
will now work also. Once this code is in the field for a while I'll
convert spells to use that instead.