cast: give user information about how many spells left to cast
No changelog entry becouse this feature is not done yet. I'll
probably revert this commit and commit it again when it will be complete.

It still has problems with triggers and dropped spells. Look at
http://bugs.sourcemage.org/show_bug.cgi?id=10022
(numbers are not exact but are still a good estimation since incidence
of problems is pretty rare and error is usualy only for 1 or 2 spells)

However for simple casts it gives nice feedback for how much
spells are left to cast.