sorcery: added as_needed optimization option
as_needed will add -Wl,--as-needed to LDFLAGS, which will cause the linker to
emit DT_NEEDED entries only for directly used libraries. This prevents programs
and libraries that only indirectly depend on some library from breaking when
that library has an soname change. A scenario which is very common with libtool
using build processes.