gaze: Optimize "gaze from" command
Do all the processing in a single AWK instance, and skip current
file if a match is found while matching whole path names or working
in non-verbose mode.

Both because it's more sensible and because of the whole-path
optimization, an incompatible change in behaviour has been introduced:
plain string search matches at path boundaries instead of the end of
the path name.