libdispel, dispel, resurrect: optionally protect files
This is the first proof of concept of preserving library files from
old spell versions. The current case for me is gcc being broken after
isl update, and hence full stop on any magic.

This only works for manual invocation of dispel/resurrect right now. The
idea is to add a variable PROTECT_FILES to DETAILS that triggers
the protection.