libunpack: Use stdin only implicitly by tar
GNU tar expects the input to be block-aligned unless it knows it's
reading from a pipe, but unpackers may not flush the output at the
expected block size, which results in read errors.

When using the -f flag, detection is based on the filename being "-".

This bug was exposed by bzip2 compiled against musl.