libmisc: fix bzip2/gzip paging for old file in stable grimoire
The file-5.23 in stable returns just 'binary' for compile logs
compressed with bzip2. This means that devel sorcery on stable
systems breaks crucial comands like `gaze compile`. So I added
this fallback that guesses bz2 and gz based on file name again,
to be removed once we are really sure sorcery always has fresh
file at hand. Or someone makes the code smarter to check file
version (as well as availability of the 'safe' locale C.UTF-8
that I'll need to tackle next … it's not available on old
installs).