libmisc: more robust debugging of parallel work
This prints out the whole debugging line on one go to reduce interference
of printouts from multiple sorcery processes (subshells for cast passes).
Debugging is a nightmare when data within a line is already interleaved!
This is still not proper, but we don't want to bring things to a crawl by
i/o locking. An alternative might be to write output to per-process files.
But then, we also should implement timestamps in the output. Pro-tip: use
ts from moreutils in a tail -f | ts configuration to put them on closely
after the fact.

Also, to actually tell the differing subshells of the cast process apart,
the variable $$ is not enough. Bash offers an alternative. This is used
with $$ as fallback now. We may want to check what to do for other shells
that we want to be able to run sorcery properly.