libdepengine: avoid building dependees after dependencies failed
This is another hack on the horrid sorcery logic of running many times
through the same code to do different things. The problem: Building of
a dependency fails, but sorcery still tries to build the spell depending
on the failed one, usually resulting in predictable failure, otherwise
in a perhaps unintended build with an older installed version of the
dependency.

Example was nss failing for me (still have to figure out that weird one)
and still firefox cast bein attempted after that. This is stupid
behaviour.

I solve that by grepping for the spell in $FAILED_LIST and using occurence
there to indicate a nonzero return value where there is no value recorded
from the depengine_cast_engine. The only justification is that it works.

Caveat: It only works to avoid nonsense in the casting phase. If things
fail in configure phase, we still have the pecuilarity that dependees are
removed from the cast only if they have been processed before. If you
specified

	cast a b

where b depends on a and configuration of a fails, b is still attempted to
be cast. This is for another night, or a proper rewrite.

Note: There are other places with grep $spell $FAILED_LIST. I suggest checking if that
should read "^$spell$" instead there, too.