libdepengine: ignore pre_remove and post_remove failures for failure detection
The loop prevention hack guessed wrongly that a pre_remove failure is
serious and hence caused lots of spells being dropped once e.g.
grep was cast, which has a PRE_REMOVE that returns non-zero (something we could
investigate). Sorcery ignores that return value also for post_remove, but
the runtime mechanism still stores the failure reason.

Maybe we should ignore that failure reason log altogether. But I think I
had a reason including it.

We need to rewrite this all ...

This commit also makes the two messages different, adding 'subtly' for the
cases where the spell didn't outright fail but still has a reason (conflicts
is one remaining case which we might want to settle) and listing the
found reasons. This mainly should aid debugging once we get rid of these
cases. Either there is a failure or not, please.